Split out CryptoHash into its own file.
