+++ linera-base/src/crypto/mod.rs
@@ -4,25 +4,18 @@
 
 //! Define the cryptographic primitives used by the Linera protocol.
 
+mod hash;
 use std::{borrow::Cow, fmt, io, num::ParseIntError, str::FromStr};
 
-use alloy_primitives::{FixedBytes, Keccak256, B256};
+use alloy_primitives::FixedBytes;
 use ed25519_dalek::{self as dalek, Signer, Verifier};
+pub use hash::*;
 use linera_witty::{
     GuestPointer, HList, InstanceWithMemory, Layout, Memory, Runtime, RuntimeError, RuntimeMemory,
     WitLoad, WitStore, WitType,
 };
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
-#[cfg(with_testing)]
-use {
-    proptest::{
-        collection::{vec, VecStrategy},
-        prelude::Arbitrary,
-        strategy::{self, Strategy},
-    },
-    std::ops::RangeInclusive,
-};
 
 use crate::doc_scalar;
 
@@ -32,21 +25,6 @@ pub struct KeyPair(dalek::SigningKey);
 /// A signature public key.
 #[derive(Eq, PartialEq, Ord, PartialOrd, Copy, Clone, Hash)]
 pub struct PublicKey(pub [u8; dalek::PUBLIC_KEY_LENGTH]);
-
-type HasherOutputSize = FixedBytes<32>;
-
-/// A Keccak256 value.
-#[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Copy, Hash)]
-#[cfg_attr(with_testing, derive(Default))]
-pub struct CryptoHash(B256);
-
-/// A vector of cryptographic hashes.
-/// This is used to represent a hash of a list of hashes.
-#[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Hash, Serialize, Deserialize)]
-#[cfg_attr(with_testing, derive(Default))]
-pub struct CryptoHashVec(pub Vec<CryptoHash>);
-
-impl<'de> BcsHashable<'de> for CryptoHashVec {}
 
 /// A signature value.
 #[derive(Eq, PartialEq, Copy, Clone)]
@@ -64,7 +42,7 @@ pub enum CryptoError {
     NonHexDigits(#[from] hex::FromHexError),
     #[error(
         "Byte slice has length {0} but a `CryptoHash` requires exactly {expected} bytes",
-        expected = HasherOutputSize::len_bytes(),
+        expected = FixedBytes::<32>::len_bytes(),
     )]
     IncorrectHashSize(usize),
     #[error(
@@ -166,39 +144,6 @@ impl<'de> Deserialize<'de> for PublicKey {
     }
 }
 
-impl Serialize for CryptoHash {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::ser::Serializer,
-    {
-        if serializer.is_human_readable() {
-            serializer.serialize_str(&self.to_string())
-        } else {
-            serializer.serialize_newtype_struct("CryptoHash", &self.as_bytes().0)
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for CryptoHash {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::de::Deserializer<'de>,
-    {
-        if deserializer.is_human_readable() {
-            let s = String::deserialize(deserializer)?;
-            let value = Self::from_str(&s).map_err(serde::de::Error::custom)?;
-            Ok(value)
-        } else {
-            #[derive(Deserialize)]
-            #[serde(rename = "CryptoHash")]
-            struct Foo([u8; 32]);
-
-            let value = Foo::deserialize(deserializer)?;
-            Ok(Self(value.0.into()))
-        }
-    }
-}
-
 impl Serialize for KeyPair {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
@@ -294,38 +239,6 @@ impl From<PublicKey> for [u64; 4] {
     }
 }
 
-impl FromStr for CryptoHash {
-    type Err = CryptoError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let value = hex::decode(s)?;
-        (value.as_slice()).try_into()
-    }
-}
-
-impl TryFrom<&[u8]> for CryptoHash {
-    type Error = CryptoError;
-
-    fn try_from(value: &[u8]) -> Result<Self, Self::Error> {
-        if value.len() != B256::len_bytes() {
-            return Err(CryptoError::IncorrectHashSize(value.len()));
-        }
-        Ok(Self(B256::from_slice(value)))
-    }
-}
-
-impl From<[u64; 4]> for CryptoHash {
-    fn from(integers: [u64; 4]) -> Self {
-        CryptoHash(u64_array_to_be_bytes(integers).into())
-    }
-}
-
-impl From<CryptoHash> for [u64; 4] {
-    fn from(crypto_hash: CryptoHash) -> Self {
-        be_bytes_to_u64_array(crypto_hash.0.as_ref())
-    }
-}
-
 impl fmt::Display for Signature {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let s = hex::encode(self.0.to_bytes());
@@ -339,13 +252,6 @@ impl fmt::Display for PublicKey {
     }
 }
 
-impl fmt::Display for CryptoHash {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let prec = f.precision().unwrap_or(self.0.len() * 2);
-        hex::encode(&self.0[..((prec + 1) / 2)]).fmt(f)
-    }
-}
-
 impl fmt::Debug for Signature {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", hex::encode(&self.0.to_bytes()[0..8]))
@@ -353,12 +259,6 @@ impl fmt::Debug for Signature {
 }
 
 impl fmt::Debug for PublicKey {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(&self.0[..8]))
-    }
-}
-
-impl fmt::Debug for CryptoHash {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", hex::encode(&self.0[..8]))
     }
@@ -401,20 +301,6 @@ where
     }
 }
 
-/// Temporary struct to extend `Keccak256` with `io::Write`.
-struct Keccak256Ext(Keccak256);
-
-impl io::Write for Keccak256Ext {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        self.0.update(buf);
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        Ok(())
-    }
-}
-
 impl<Hasher> Hashable<Hasher> for [u8]
 where
     Hasher: io::Write,
@@ -430,26 +316,6 @@ where
 {
     fn type_name() -> &'static str {
         serde_name::trace_name::<Self>().expect("Self must be a struct or an enum")
-    }
-}
-
-impl CryptoHash {
-    /// Computes a hash.
-    pub fn new<'de, T: BcsHashable<'de>>(value: &T) -> Self {
-        let mut hasher = Keccak256Ext(Keccak256::new());
-        value.write(&mut hasher);
-        CryptoHash(hasher.0.finalize())
-    }
-
-    /// Reads the bytes of the hash value.
-    pub fn as_bytes(&self) -> &B256 {
-        &self.0
-    }
-
-    /// Returns the hash of `TestString(s)`, for testing purposes.
-    #[cfg(with_testing)]
-    pub fn test_hash(s: impl Into<String>) -> Self {
-        CryptoHash::new(&TestString::new(s))
     }
 }
 
@@ -544,81 +410,6 @@ impl Signature {
     }
 }
 
-impl WitType for CryptoHash {
-    const SIZE: u32 = <(u64, u64, u64, u64) as WitType>::SIZE;
-    type Layout = <(u64, u64, u64, u64) as WitType>::Layout;
-    type Dependencies = HList![];
-
-    fn wit_type_name() -> Cow<'static, str> {
-        "crypto-hash".into()
-    }
-
-    fn wit_type_declaration() -> Cow<'static, str> {
-        concat!(
-            "    record crypto-hash {\n",
-            "        part1: u64,\n",
-            "        part2: u64,\n",
-            "        part3: u64,\n",
-            "        part4: u64,\n",
-            "    }\n",
-        )
-        .into()
-    }
-}
-
-impl WitLoad for CryptoHash {
-    fn load<Instance>(
-        memory: &Memory<'_, Instance>,
-        location: GuestPointer,
-    ) -> Result<Self, RuntimeError>
-    where
-        Instance: InstanceWithMemory,
-        <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>,
-    {
-        let (part1, part2, part3, part4) = WitLoad::load(memory, location)?;
-        Ok(CryptoHash::from([part1, part2, part3, part4]))
-    }
-
-    fn lift_from<Instance>(
-        flat_layout: <Self::Layout as Layout>::Flat,
-        memory: &Memory<'_, Instance>,
-    ) -> Result<Self, RuntimeError>
-    where
-        Instance: InstanceWithMemory,
-        <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>,
-    {
-        let (part1, part2, part3, part4) = WitLoad::lift_from(flat_layout, memory)?;
-        Ok(CryptoHash::from([part1, part2, part3, part4]))
-    }
-}
-
-impl WitStore for CryptoHash {
-    fn store<Instance>(
-        &self,
-        memory: &mut Memory<'_, Instance>,
-        location: GuestPointer,
-    ) -> Result<(), RuntimeError>
-    where
-        Instance: InstanceWithMemory,
-        <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>,
-    {
-        let [part1, part2, part3, part4] = (*self).into();
-        (part1, part2, part3, part4).store(memory, location)
-    }
-
-    fn lower<Instance>(
-        &self,
-        memory: &mut Memory<'_, Instance>,
-    ) -> Result<<Self::Layout as Layout>::Flat, RuntimeError>
-    where
-        Instance: InstanceWithMemory,
-        <Instance::Runtime as Runtime>::Memory: RuntimeMemory<Instance>,
-    {
-        let [part1, part2, part3, part4] = (*self).into();
-        (part1, part2, part3, part4).lower(memory)
-    }
-}
-
 impl WitType for PublicKey {
     const SIZE: u32 = <(u64, u64, u64, u64) as WitType>::SIZE;
     type Layout = <(u64, u64, u64, u64) as WitType>::Layout;
@@ -694,20 +485,8 @@ impl WitStore for PublicKey {
     }
 }
 
-#[cfg(with_testing)]
-impl Arbitrary for CryptoHash {
-    type Parameters = ();
-    type Strategy = strategy::Map<VecStrategy<RangeInclusive<u8>>, fn(Vec<u8>) -> CryptoHash>;
-
-    fn arbitrary_with((): Self::Parameters) -> Self::Strategy {
-        vec(u8::MIN..=u8::MAX, HasherOutputSize::len_bytes())
-            .prop_map(|vector| CryptoHash(B256::from_slice(&vector[..])))
-    }
-}
-
 impl<'de> BcsHashable<'de> for PublicKey {}
 
-doc_scalar!(CryptoHash, "A Keccak256 value");
 doc_scalar!(PublicKey, "A signature public key");
 doc_scalar!(Signature, "A signature value");
 
@@ -751,7 +530,7 @@ fn test_signatures() {
 }
 
 /// Reads the `bytes` as four little-endian unsigned 64-bit integers and returns them.
-fn le_bytes_to_u64_array(bytes: &[u8]) -> [u64; 4] {
+pub(crate) fn le_bytes_to_u64_array(bytes: &[u8]) -> [u64; 4] {
     let mut integers = [0u64; 4];
 
     integers[0] = u64::from_le_bytes(bytes[0..8].try_into().expect("incorrect indices"));
@@ -763,7 +542,7 @@ fn le_bytes_to_u64_array(bytes: &[u8]) -> [u64; 4] {
 }
 
 /// Reads the `bytes` as four big-endian unsigned 64-bit integers and returns them.
-fn be_bytes_to_u64_array(bytes: &[u8]) -> [u64; 4] {
+pub(crate) fn be_bytes_to_u64_array(bytes: &[u8]) -> [u64; 4] {
     let mut integers = [0u64; 4];
 
     integers[0] = u64::from_be_bytes(bytes[0..8].try_into().expect("incorrect indices"));
@@ -775,7 +554,7 @@ fn be_bytes_to_u64_array(bytes: &[u8]) -> [u64; 4] {
 }
 
 /// Returns the bytes that represent the `integers` in little-endian.
-fn u64_array_to_le_bytes(integers: [u64; 4]) -> [u8; 32] {
+pub(crate) fn u64_array_to_le_bytes(integers: [u64; 4]) -> [u8; 32] {
     let mut bytes = [0u8; 32];
 
     bytes[0..8].copy_from_slice(&integers[0].to_le_bytes());
@@ -787,7 +566,7 @@ fn u64_array_to_le_bytes(integers: [u64; 4]) -> [u8; 32] {
 }
 
 /// Returns the bytes that represent the `integers` in big-endian.
-fn u64_array_to_be_bytes(integers: [u64; 4]) -> [u8; 32] {
+pub(crate) fn u64_array_to_be_bytes(integers: [u64; 4]) -> [u8; 32] {
     let mut bytes = [0u8; 32];
 
     bytes[0..8].copy_from_slice(&integers[0].to_be_bytes());
